maked holder, adapter and layout
