--- app/src/main/java/com/edurda77/dictionary/MainActivity.kt
@@ -1,14 +1,0 @@
-package com.edurda77.dictionary
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.edurda77.dictionary.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-    }
-}
